feat: add support for new architecture on iOS
