+++ ios/Adapters/BearerAuthenticationAdapter.swift
@@ -1,12 +1,3 @@
-//
-//  BearerAuthenticationAdapter.swift
-//  NetworkClient
-//
-//  Created by Miguel Alatzar on 12/2/20.
-//  Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
-//  See LICENSE.txt for license information.
-//
-
 import Foundation
 import Alamofire
