+++ ios/Extensions/Session+Extensions.swift
@@ -1,12 +1,3 @@
-//
-//  Session+Extensions.swift
-//  NetworkClient
-//
-//  Created by Miguel Alatzar on 11/18/20.
-//  Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
-//  See LICENSE.txt for license information.
-//
-
 import Alamofire
 
 extension Session: Equatable {
