+++ ios/Extensions/StringProtocol+Extensions.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+extension StringProtocol {
+    var firstUppercased: String { return prefix(1).uppercased() + dropFirst() }
+}
