+++ ios/Extensions/URL+Extensions.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+extension URL {
+    func fileSize() throws -> Int {
+        return try self.resourceValues(forKeys: [.fileSizeKey]).fileSize!
+    }
+}
