+++ ios/Extensions/URLRequest+Extensions.swift
@@ -1,12 +1,3 @@
-//
-//  URLRequest+Extensions.swift
-//  NetworkClient
-//
-//  Created by Miguel Alatzar on 4/23/21.
-//  Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
-//  See LICENSE.txt for license information.
-//
-
 import Foundation
 import Alamofire
