+++ ios/NetworkClient-Bridging-Header.h
@@ -1,18 +1,1 @@
-//
-//  NetworkClient-Bridging-Header.h
-//  NetworkClient
-//
-//  Created by Miguel Alatzar on 12/2/20.
-//  Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
-//  See LICENSE.txt for license information.
-//
 
-#ifndef NetworkClient_Bridging_Header_h
-#define NetworkClient_Bridging_Header_h
-
-#import <React/RCTBridge.h>
-#import <React/RCTEventDispatcher.h>
-#import <React/RCTBridgeModule.h>
-#import <React/RCTConvert.h>
-
-#endif /* NetworkClient_Bridging_Header_h */
