+++ ios/SessionManager.swift
@@ -1,12 +1,3 @@
-//
-//  SessionManager.swift
-//  NetworkClient
-//
-//  Created by Miguel Alatzar on 9/10/20.
-//  Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
-//  See LICENSE.txt for license information.
-//
-
 import Foundation
 import Alamofire
 import SwiftyJSON
