--- ios/WebSocketClient.m
@@ -1,30 +1,0 @@
-//
-//  WebSocketClient.m
-//  NetworkClient
-//
-//  Created by Miguel Alatzar on 1/18/21.
-//  Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
-//  See LICENSE.txt for license information.
-//
-
-#import <React/RCTBridgeModule.h>
-#import <React/RCTEventEmitter.h>
-
-@interface RCT_EXTERN_MODULE(WebSocketClient, RCTEventEmitter)
-
-RCT_EXTERN_METHOD(supportedEvents)
-
-RCT_EXTERN_METHOD(ensureClientFor:(NSString *)url withOptions:(NSDictionary *)options withResolver:(RCTPromiseResolveBlock)resolve withRejecter:(RCTPromiseRejectBlock)reject)
-
-RCT_EXTERN_METHOD(createClientFor:(NSString *)url withOptions:(NSDictionary *)options withResolver:(RCTPromiseResolveBlock)resolve withRejecter:(RCTPromiseRejectBlock)reject)
-
-RCT_EXTERN_METHOD(connectFor:(NSString *)url withResolver:(RCTPromiseResolveBlock)resolve withRejecter:(RCTPromiseRejectBlock)reject)
-
-RCT_EXTERN_METHOD(disconnectFor:(NSString *)url withResolver:(RCTPromiseResolveBlock)resolve withRejecter:(RCTPromiseRejectBlock)reject)
-
-RCT_EXTERN_METHOD(sendDataFor:(NSString *)url withData:(NSString *)data withResolver:(RCTPromiseResolveBlock)resolve withRejecter:(RCTPromiseRejectBlock)reject)
-
-RCT_EXTERN_METHOD(invalidateClientFor:(NSString *)url withResolver:(RCTPromiseResolveBlock)resolve withRejecter:(RCTPromiseRejectBlock)reject)
-
-
-@end
